Restructuring so that eventually utils.py can be called wrappers.py
