+++ wllvm/wllvmpp.py
@@ -8,10 +8,13 @@ file so that it can be found later after all of the objects are linked
 into a library or executable.
 """
 
-import sys
+import sys, os
 
-from utils import *
+from utils import getBuilder, buildObject, buildAndAttachBitcode, logging
+
 import logconfig
+
+_logger = logging.getLogger(__name__)
 
 def main():
     try:
@@ -20,9 +23,12 @@ def main():
         cmd = cmd[1:]
         builder = getBuilder(cmd, True)
         rc = buildObject(builder)
+
+ 
         if rc == 0 and not os.environ.get('WLLVM_CONFIGURE_ONLY', False):
             buildAndAttachBitcode(builder)
-    except:
+    except Exception as e:
+        _logger.debug('wllvm++: exception case: {0}\n'.format(e))
         pass
     return rc
